refactor: move common components to @/components folder
